% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIRE Data Set}
\format{
A data frame with 15 rows and 2 variables
\describe{
 \item{DISTANCE}{The distance from the fire station (Miles)}
 \item{DAMAGE}{Amt of damage (In 1000s of $)}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set tracking home damage from fires by measuring the distance a fire was from the
fire station as well as the damage the fire caused
}
\keyword{datasets}
