% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Used to make bootstrap samples on a specific population for a specific function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5)
}
\arguments{
\item{iter}{-> Number of Samples}

\item{x}{-> Population to sample from}

\item{fun}{-> Function to be performed on samples}

\item{alpha}{-> Confidence Level (1 - alpha)}

\item{cx}{-> Size of Points}
}
\value{
A histogram of the resulting function and list of confidence interval, function, and population
}
\description{
Used to make bootstrap samples on a specific population for a specific function
}
\examples{
myboot2(x = c(1, 2, 3, 4, 5))
}
