% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{The "My Central Limit Theorem - Uniform" function: Creates a histogram of sample means from uniform distribution from a to b}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{-> Size of each sample}

\item{iter}{-> Number of Samples}

\item{a}{-> Left endpoint of uniform distribution}

\item{b}{-> Right endpoint of uniform distribution}
}
\value{
A histogram of all the samples means, with a theoretical normal curve
}
\description{
The "My Central Limit Theorem - Uniform" function: Creates a histogram of sample means from uniform distribution from a to b
}
\examples{
mycltu(10, 1000, 0, 5)
}
