% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{The mymlnorm function perform a maximum likelihood estimate for a normal experiment}
\usage{
mymlnorm(x, mu, sig)
}
\arguments{
\item{x}{-> Sample Data}

\item{mu}{-> Possible values of mu}

\item{sig}{-> Possible values of sigma}
}
\value{
A plot with maximum likelihood estimates graphically shown
}
\description{
The mymlnorm function perform a maximum likelihood estimate for a normal experiment
}
\examples{
mymlnorm(x = c(12, 13, 15), mu = seq(12, 15, length.out = 100), sig = seq(1, 5, length.out = 100))
}
