% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve Function -> Displays a normal distribution curve with shaded probability P(Y <= a)}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{-> Mean of normal distribution}

\item{sigma}{-> Standard deviation of normal distribution}

\item{a}{-> Shades probability of normal distribution P(Y <= a)}
}
\value{
A normal distribution curve with the shaded area and a list with mu, sigma, and probability
}
\description{
myncurve Function -> Displays a normal distribution curve with shaded probability P(Y <= a)
}
\examples{
myncurve(10, 10, 9)
}
