% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysummary.R
\name{mysummary}
\alias{mysummary}
\title{My Summary Function}
\usage{
mysummary(x)
}
\arguments{
\item{x}{-> A regression model object}
}
\value{
A formatted summary of the object with the data we want
}
\description{
My Summary Function
}
