% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets Function -> Used to solve the plane overbooking problem}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.5)
}
\arguments{
\item{N}{-> Number of seats on the plane}

\item{gamma}{-> Probability of overbooking}

\item{p}{-> Probability a given passenger will show up to the flight}
}
\value{
A named list of both the discrete and continuous results for the optimals number of tickets sold, as well as 2 plots showing the objective vs. n for both the discrete and continuous cases
}
\description{
ntickets Function -> Used to solve the plane overbooking problem
}
\examples{
ntickets(400, 0.02, 0.95)
}
