% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysquare.R
\name{mysquare}
\alias{mysquare}
\title{The Square Function squares each value passed to it}
\usage{
mysquare(x)
}
\arguments{
\item{x}{-> A vector of quantitative Data}
}
\value{
a Vector of squared values
}
\description{
The Square Function squares each value passed to it
}
\examples{
mysquare(x = 1:10)
}
